requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient 
gunicorn  # For production (optional)
